requirements.txt
amqp==2.6.0
arrow==0.15.8
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
click==7.1.2
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
install==1.3.3
kombu==4.6.11
Markdown==3.2.2
pdfcrowd==4.12.0
pdfkit==0.6.1
proj==0.2.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
